retriever setup but disambiguity might give it a tough time
